Add files via upload
This project has Experiment which is a unit of measurable research that defines a single run with some data, parameters, code, or results. We will first load our EPL 2020-21 dataset as a pandas dataframe using read_csv(). We will send our experiment to an already existing project and workspace in comet.ml. If project name does not exist, comet.ml will create a new project.